requirements.txt
Django==3.1.2
psycopg2==2.8.6
pytest-django==4.0.0
